+++ packages/moonstone/ProgressBar/ProgressBar.js
@@ -18,6 +18,7 @@ import compose from 'ramda/src/compose';
 import React from 'react';
 
 import Skinnable from '../Skinnable';
+import {ProgressBarTooltip} from './ProgressBarTooltip';
 
 import componentCss from './ProgressBar.less';
 
@@ -44,7 +45,69 @@ const ProgressBarBase = kind({
 		 * @type {Object}
 		 * @public
 		 */
-		css: PropTypes.object
+		css: PropTypes.object,
+
+		/**
+		 * Sets the orientation of the slider, whether the progress-bar depicts its progress value
+		 * in a left and right orientation or up and down onientation.
+		 * Must be either `'horizontal'` or `'vertical'`.
+		 *
+		 * @type {String}
+		 * @default 'horizontal'
+		 * @public
+		 */
+		orientation: PropTypes.oneOf(['horizontal', 'vertical']),
+
+		/**
+		 * The proportion of the filled portion of the progress bar. Valid values are
+		 * between `0` and `1`.
+		 *
+		 * @type {Number}
+		 * @default 0
+		 * @public
+		 */
+		progress: PropTypes.number,
+
+		/**
+		 * Enables the built-in tooltip, whose behavior can be modified by the other tooltip
+		 * properties.
+		 *
+		 * @type {Boolean}
+		 * @public
+		 */
+		tooltip: PropTypes.bool,
+
+		/**
+		 * Setting to `true` overrides the natural tooltip position
+		 * for `vertical` progress bar to the left. This may be useful
+		 * if you have a static layout that have the progress bar at
+		 * the right edge of the container.
+		 *
+		 * @type {Boolean}
+		 * @public
+		 */
+		tooltipForceSide: PropTypes.bool,
+
+		/**
+		 * Specify where the tooltip should appear in relation to the ProgressBar. Options are
+		 * `'before'` and `'after'`. `before` renders above a `horizontal` slider and to the
+		 * left of a `vertical` ProgressBar. `after` renders below a `horizontal` slider and to the
+		 * right of a `vertical` ProgressBar. In the `vertical` case, the rendering position is
+		 * automatically reversed when rendering in an RTL locale. This can be overridden by
+		 * using the [tooltipForceSide]{@link moonstone/ProgressBar.ProgressBar.tooltipForceSide}
+		 * prop.
+		 *
+		 * @type {String}
+		 * @default 'before'
+		 * @public
+		 */
+		tooltipSide: PropTypes.string
+	},
+
+	defaultProps: {
+		tooltip: false,
+		tooltipSide: 'before',
+		tooltipForceSide: false
 	},
 
 	styles: {
@@ -52,12 +115,42 @@ const ProgressBarBase = kind({
 		publicClassNames: ['progressBar']
 	},
 
-	render: (props) => {
+	computed: {
+		tooltipComponent: ({orientation, progress, tooltip, tooltipForceSide, tooltipSide}) => {
+			if (tooltip) {
+				const progressAfterMidpoint = progress > 0.5;
+				const progressPercentage = Math.min(parseInt(progress * 100), 100);
+				const percentageText = `${progressPercentage}%`;
+
+				return (
+					<ProgressBarTooltip
+						forceSide={tooltipForceSide}
+						knobAfterMidpoint={progressAfterMidpoint}
+						proportion={progress}
+						orientation={orientation}
+						side={tooltipSide}
+					>
+						{percentageText}
+					</ProgressBarTooltip>
+				);
+			} else {
+				return null;
+			}
+		}
+	},
+
+	render: ({css, tooltipComponent, ...rest}) => {
+		delete rest.tooltip;
+		delete rest.tooltipSide;
+		delete rest.tooltipForceSide;
+
 		return (
 			<UiProgressBar
-				{...props}
-				css={props.css}
-			/>
+				{...rest}
+				css={css}
+			>
+				{tooltipComponent}
+			</UiProgressBar>
 		);
 	}
 });
@@ -92,5 +185,6 @@ export default ProgressBar;
 export {
 	ProgressBar,
 	ProgressBarBase,
-	ProgressBarDecorator
+	ProgressBarDecorator,
+	ProgressBarTooltip
 };