+++ packages/sampler/stories/qa-stories/ProgressBar.js
@@ -1,7 +1,7 @@
 import ProgressBar, {ProgressBarBase} from '@enact/moonstone/ProgressBar';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
-import {boolean, number} from '@storybook/addon-knobs';
+import {boolean, number, select} from '@storybook/addon-knobs';
 
 import nullify from '../../src/utils/nullify.js';
 import {mergeComponentMetadata} from '../../src/utils/propTables';
@@ -13,9 +13,11 @@ storiesOf('ProgressBar', module)
 		'The basic ProgressBar',
 		() => (
 			<ProgressBar
-				backgroundProgress={number('backgroundProgress', ProgressBarBase.defaultProps.backgroundProgress, {range: true, min: 0, max: 1, step: 0.01})}
-				progress={number('progress', ProgressBarBase.defaultProps.progress, {range: true, min: 0, max: 1, step: 0.1})}
-				vertical={boolean('vertical', false)}
+				backgroundProgress={number('backgroundProgress', 0.5, {range: true, min: 0, max: 1, step: 0.01})}
+				tooltip={boolean('tooltip', false)}
+				progress={number('progress', 0.4, {range: true, min: 0, max: 1, step: 0.01})}
+				orientation={select('orientation', ['horizontal', 'vertical'], 'horizontal')}
+				tooltipForceSide={boolean('tooltipForceSide', false)}
 				disabled={nullify(boolean('disabled', false))}
 			/>
 		),