+++ packages/ui/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The following is a curated list of changes in the Enact ui module, newest changes on the top.
 
+## [unreleased]
+
+### Changed
+
+- `ui/Transition` property `children` not to be required
+- `ui/Transition` to fire `onShow` and `onHide` even when there are no `children`
+
 ## [2.0.0-alpha.6] - 2018-03-22
 
 ### Removed